Update README.md
I was having a relaxing night, so I decided to offer a few copywriting tweaks to tighten wording and correct typos. I hope it helps! One thing: It’s unclear to me whether “Uses openssl with the aes-256-cbc ciper” *[sic]* is a *STEP* in the encryption process or just a *NOTE* about secure data transmission. If it’s the *former* you might explain *WHAT* it does. If it’s the latter, you might consider moving it to the end of the section by itself as a final section note.